backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
openai>=1.0.0
duckduckgo-search>=4.0.0
tavily-python>=0.3.0
